pyproject.toml
[tool.poetry]
name = "app-ml-fastapi"
version = "0.1.0"
description = ""
authors = ["Botalov Sergey <89090168690@mail.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
